Close closed issues of target repository
Under existing behaviour closed issues are imported as open issues with a title beginning [CLOSED].
This change closes imported issues in the target repo if they were closed in the source repo.

It also identifies the number of open and closed issues when asking the user whether they wish to continue.

You are about to add to 'AcmeCorp/projectX':
 * 60 open issues
 * 934 closed issues
 * 0 pull requests
 * 3248 new comments
 * 31 new milestones
 * 83 new labels
Are you sure you wish to continue? [Y/n]

* Contributor: https://github.com/rosscado

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>