Added --ignore-pull-requests option
This commit is to append --ignore-pull-requests to do not run
the import operation of pull requests.

The existing version imports the numbers of pull requests as well as
that of issues in to an issue list of a target repository.

**Changes proposed in this PR:**
1. Added an option to skip an import procedure of pull requests
2. Added --ignore-pull-requests by default

* After this PR:
If this option is specified then pull request type issues are not imported.
--ignore-pull-requests   Do not import pull requests.

* Author:
https://github.com/rosscado

Signed-off-by: Geunsik Lim <leemgs@gmail.com>