Added variable in config file for controlling issue numbers of group
After this PR, users can declare the number of issues that you want to maintain
as a group to avoid a service denial when webhook API requests exceed a max number
of a rate limit in github.com.

Note: If you meet "ERROR: There was a problem during authentication", we recommend
that you try to define smaller numbers instead of 20 (binding issues as a group).
For example, 20(default), 10, 8, 6, 4, and 2 issues.
When the number of specified issues are proceeded, the program sleep for a specified time
before starting processing a next issue group.

* User case: ISSUE_ITEMS_GROUP=20

Signed-off-by: Geunsik Lim <geunsik.lim@samsung.com>